# backend/api.yaml
openapi: 3.0.2
servers:
  - url: https://gigger-api.orionmmtecheng.com
info:
  description: This is API Doc for Gigger
  title: Swagger API Doc for Gigger
  version: 1.0.0
...
paths:
...
  /api/logout:
    get:
      tags:
        - Authentication
      summary: Logs out current logged in user session
      security:
        - {}
      description: ""
      operationId: logoutUser
      parameters:
        - name: userId
          in: header
          description: ID of login UserId
          required: true
          schema:
            type: string
      responses:
        default:
          description: Logout Successfully.
...
components:
...
  securitySchemes:
    bearerAuth: # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT # optional, arbitrary value for documentation purposes
...
security:
  - bearerAuth: []